修复 PR Review。
- 为 attachment 表添加索引。
- FsHttpEndpoint 通过方法获取 FsService，避免 FsService 不存在的情况。
- 删除 S3UploadId 类的 final 修饰符。
- 优化 FS 大文件合并，避免 OOM。
- 将文件大小类型改为 long。
- 补充注释。

+++ zeroweb-service/zeroweb-service-file/src/main/java/io/github/xezzon/zeroweb/attachment/entity/UploadInfo.java
@@ -12,6 +12,6 @@ public record UploadInfo(
     String id,
     FileProviderEnum provider,
     int partCount,
-    int partSize
+    long partSize
 ) {
 }
